+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+Migrated to Typescript 5.4
+
+### Removed
+`members` field for the API metadata. All its functionalities are in `implementation` in a more type-safe way
 
 ## 2.4.1 - 2024-03-27
 ### Added